Add Cas1AssessmentService
- Added cas1 assessment service
- Added cas1 assessment transformer
- Removed cas1 assessment functions from legacy code and tests.
