+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/transformer/cas1/Cas1AssessmentClarificationNoteTransformer.kt
@@ -1,13 +1,12 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.cas1
 
 import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ClarificationNote
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1ClarificationNote
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AssessmentClarificationNoteEntity
 
 @Component
-@Deprecated("Will be removed soon - use cas specific version instead")
-class AssessmentClarificationNoteTransformer {
-  fun transformJpaToApi(jpa: AssessmentClarificationNoteEntity) = ClarificationNote(
+class Cas1AssessmentClarificationNoteTransformer {
+  fun transformJpaToCas1ClarificationNote(jpa: AssessmentClarificationNoteEntity) = Cas1ClarificationNote(
     id = jpa.id,
     createdAt = jpa.createdAt.toInstant(),
     createdByStaffMemberId = jpa.createdByUser.id,
