+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/transformer/AssessmentTransformerTest.kt
@@ -24,7 +24,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremis
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremisesAssessmentSummary
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremisesUser
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1Application
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1Assessment
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1AssessmentStatus
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1AssessmentSummary
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Person
@@ -397,138 +396,6 @@ class AssessmentTransformerTest {
           },
         ),
       )
-    }
-  }
-
-  @Nested
-  inner class TransformJpaToCas1Assessment {
-
-    @Test
-    fun `transformJpaToApi transforms correctly`() {
-      val assessment = approvedPremisesAssessmentFactory
-        .withData("{ \"some\": \"data\" }")
-        .withDocument("{ \"some\": \"doc\" }")
-        .produce()
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result.id).isEqualTo(UUID.fromString("7d0d3b38-5bc3-45c7-95eb-4d714cbd0db1"))
-      assertThat(result.schemaVersion).isEqualTo(UUID.fromString("aeeb6992-6485-4600-9c35-19479819c544"))
-      assertThat(result.decision).isEqualTo(ApiAssessmentDecision.rejected)
-      assertThat(result.rejectionRationale).isEqualTo("reasoning")
-      assertThat(result.createdAt).isEqualTo(Instant.parse("2022-12-14T12:05:00Z"))
-      assertThat(result.submittedAt).isEqualTo(Instant.parse("2022-12-14T12:06:00Z"))
-      assertThat(result.allocatedToStaffMember).isEqualTo(approvedPremisesUser)
-      assertThat(result.createdFromAppeal).isEqualTo(assessment.createdFromAppeal)
-      assertThat(result.data.toString()).isEqualTo("{\"some\":\"data\"}")
-      assertThat(result.document.toString()).isEqualTo("{\"some\":\"doc\"}")
-
-      verify { mockUserTransformer.transformCas1JpaToApi(allocatedToUser) }
-    }
-
-    @Test
-    fun `transformJpaToApi for CAS1 handles minimum null values`() {
-      val assessment = approvedPremisesAssessmentFactory
-        .withAllocatedAt(null)
-        .withAllocatedToUser(null)
-        .produce()
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result.id).isEqualTo(UUID.fromString("7d0d3b38-5bc3-45c7-95eb-4d714cbd0db1"))
-      assertThat(result.allocatedAt).isNull()
-      assertThat(result.allocatedToStaffMember).isNull()
-    }
-
-    @Test
-    fun `transformJpaToApi for Approved Premises sets a pending status when there is a clarification note with no response`() {
-      val assessment = approvedPremisesAssessmentFactory.withDecision(null).produce()
-
-      val clarificationNotes = mutableListOf(
-        AssessmentClarificationNoteEntityFactory()
-          .withAssessment(assessment)
-          .withResponse("Some text")
-          .withCreatedBy(
-            UserEntityFactory()
-              .withYieldedProbationRegion {
-                ProbationRegionEntityFactory()
-                  .withYieldedApArea { ApAreaEntityFactory().produce() }
-                  .produce()
-              }
-              .produce(),
-          )
-          .produce(),
-        AssessmentClarificationNoteEntityFactory()
-          .withAssessment(assessment)
-          .withCreatedBy(
-            UserEntityFactory()
-              .withYieldedProbationRegion {
-                ProbationRegionEntityFactory()
-                  .withYieldedApArea { ApAreaEntityFactory().produce() }
-                  .produce()
-              }
-              .produce(),
-          )
-          .produce(),
-      )
-
-      assessment.clarificationNotes = clarificationNotes
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result).isInstanceOf(Cas1Assessment::class.java)
-      assertThat(result.status).isEqualTo(Cas1AssessmentStatus.awaitingResponse)
-    }
-
-    @Test
-    fun `transformJpaToApi for Approved Premises sets a completed status when there is a decision`() {
-      val assessment = approvedPremisesAssessmentFactory
-        .withDecision(JpaAssessmentDecision.ACCEPTED)
-        .produce()
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result).isInstanceOf(Cas1Assessment::class.java)
-      assertThat(result.status).isEqualTo(Cas1AssessmentStatus.completed)
-    }
-
-    @Test
-    fun `transformJpaToApi for Approved Premises sets a deallocated status when there is a deallocated timestamp`() {
-      val assessment = approvedPremisesAssessmentFactory
-        .withDecision(null)
-        .withReallocatedAt(OffsetDateTime.now())
-        .produce()
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result).isInstanceOf(Cas1Assessment::class.java)
-      assertThat(result.status).isEqualTo(Cas1AssessmentStatus.reallocated)
-    }
-
-    @Test
-    fun `transformJpaToApi for Approved Premises sets an inProgress status when there is no decision and the assessment has data`() {
-      val assessment = approvedPremisesAssessmentFactory
-        .withData("{\"data\": \"something\"}")
-        .withDecision(null)
-        .produce()
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result).isInstanceOf(Cas1Assessment::class.java)
-      assertThat(result.status).isEqualTo(Cas1AssessmentStatus.inProgress)
-    }
-
-    @Test
-    fun `transformJpaToApi for Approved Premises sets a notStarted status when there is no decision and the assessment has no data`() {
-      val assessment = approvedPremisesAssessmentFactory
-        .withData(null)
-        .withDecision(null)
-        .produce()
-
-      val result = assessmentTransformer.transformJpaToCas1Assessment(assessment, mockk())
-
-      assertThat(result).isInstanceOf(Cas1Assessment::class.java)
-      assertThat(result.status).isEqualTo(Cas1AssessmentStatus.notStarted)
     }
   }
